feat: update priority attribute to int64 and add validation for incrementing sequences in smart computer and mobile device groups

+++ internal/services/smart_computer_group_v2/data_source_read.go
@@ -108,7 +108,7 @@ func (d *smartComputerGroupV2FrameworkDataSource) Read(ctx context.Context, req 
 	for _, criterion := range resource.Criteria {
 		criteriaModel := smartComputerGroupV2CriteriaDataModel{
 			Name:       types.StringValue(criterion.Name),
-			Priority:   types.Int32Value(int32(criterion.Priority)),
+			Priority:   types.Int64Value(int64(criterion.Priority)),
 			AndOr:      types.StringValue(strings.ToLower(criterion.AndOr)),
 			SearchType: types.StringValue(criterion.SearchType),
 			Value:      types.StringValue(criterion.Value),
